requirements.txt
streamlit
pandas
plotly
numpy
openpyxl
scikit-learn
openai
python-dotenv